+++ R/bru.inference.R
@@ -269,7 +269,9 @@ single_stackmaker <- function(model, lhood, result) {
       make.stack(
         data = lhood$ips, model = model, expr = lhood$expr, y = 0,
         E = lhood$E * lhood$ips$weight, offset = 0, result = result
-      )
+      ),
+      # Make sure components with zero derivative are kept:
+      remove.unused = FALSE
     )
   } else {
     make.stack(
@@ -283,11 +285,15 @@ single_stackmaker <- function(model, lhood, result) {
 joint_stackmaker <- function(model, lhoods, result) {
   do.call(
     inla.stack.mjoin,
-    lapply(
-      lhoods,
-      function(lh) {
-        single_stackmaker(model, lh, result)
-      }
+    c(
+      lapply(
+        lhoods,
+        function(lh) {
+          single_stackmaker(model, lh, result)
+        }
+      ),
+      # Make sure components with zero derivative are kept:
+      remove.unused = FALSE
     )
   )
 }
@@ -631,7 +637,10 @@ summary.bru <- function(object, ...) {
 #' @param data A data.frame or SpatialPointsDataFrame of covariates needed for the prediction.
 #' @param formula A formula determining which effects to predict and how to combine them.
 #' @param n.samples Integer setting the number of samples to draw in order to calculate the posterior statistics. The default is rather low but provides a quick approximate result.
-#' @param seed Random number genreator seed passed on to \code{inla.posterior.sample}
+#' @param seed Random number generator seed passed on to \code{inla.posterior.sample}
+#' @param num.threads Specification of desired number of threads for parallel
+#' computations. Default NULL, leaves it up to INLA.
+#' When seed != 0, overridden to "1:1"
 #' @param ... ignored arguments (S3 generic compatibility).
 #'
 #' @return a data.frame or Spatial* object with predicted mean values and other summary statistics attached.
@@ -642,6 +651,7 @@ predict.bru <- function(object,
                         formula = NULL,
                         n.samples = 100,
                         seed = 0L,
+                        num.threads = NULL,
                         ...) {
 
   # Convert data into list, data.frame or a Spatial object if not provided as such
@@ -652,7 +662,12 @@ predict.bru <- function(object,
     data <- vertices(data)
   }
 
-  vals <- generate.bru(object, data, formula = formula, n.samples = n.samples, seed = seed)
+  vals <- generate.bru(object,
+                       data,
+                       formula = formula,
+                       n.samples = n.samples,
+                       seed = seed,
+                       num.threads = num.threads)
 
   # Summarize
 
@@ -700,7 +715,10 @@ predict.bru <- function(object,
 #' @param formula A formula determining which effects to sample from and how to combine them analytically.
 #' @param n.samples Integer setting the number of samples to draw in order to calculate the posterior statistics.
 #'                  The default is rather low but provides a quick approximate result.
-#' @param seed Random number genreator seed passed on to \code{inla.posterior.sample}
+#' @param seed Random number generator seed passed on to \code{inla.posterior.sample}
+#' @param num.threads Specification of desired number of threads for parallel
+#' computations. Default NULL, leaves it up to INLA.
+#' When seed != 0, overridden to "1:1"
 #' @param ... additional, unused arguments.
 #'
 #' @return List of generated samples
@@ -713,6 +731,7 @@ generate.bru <- function(object,
                          formula = NULL,
                          n.samples = 100,
                          seed = 0L,
+                         num.threads = NULL,
                          ...) {
   # Convert data into list, data.frame or a Spatial object if not provided as such
   if (is.character(data)) {